requirements.txt
# A4図面解析システム 必要パッケージ

# OpenAI API
openai>=1.3.0

# Streamlit UI
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# データ処理
pandas>=1.5.0
numpy>=1.24.0

# 画像処理
opencv-python>=4.8.0
Pillow>=9.5.0
pdf2image>=1.16.0

# エクセル処理
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# 設定・データベース
PyYAML>=6.0
sqlite3

# 機械学習・数値計算
scikit-learn>=1.3.0
scipy>=1.10.0

# 可視化
plotly>=5.15.0
matplotlib>=3.7.0

# その他ユーティリティ
python-dateutil>=2.8.2
pathlib2>=2.3.7
tqdm>=4.65.0

# ログ・監視
schedule>=1.2.0

# 開発・テスト用
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0